+++ component/block_cache/block_cache.go
@@ -70,6 +70,7 @@ type BlockCache struct {
 
 	blockSize       uint64          // Size of each block to be cached
 	memSize         uint64          // Mem size to be used for caching at the startup
+	mntPath         string          // Mount path
 	tmpPath         string          // Disk path where these blocks will be cached
 	diskSize        uint64          // Size of disk space allocated for the caching
 	diskTimeout     uint32          // Timeout for which disk blocks will be cached
@@ -83,9 +84,9 @@ type BlockCache struct {
 	maxDiskUsageHit bool            // Flag to indicate if we have hit max disk usage
 	noPrefetch      bool            // Flag to indicate if prefetch is disabled
 	prefetchOnOpen  bool            // Start prefetching on file open call instead of waiting for first read
-
-	lazyWrite    bool           // Flag to indicate if lazy write is enabled
-	fileCloseOpt sync.WaitGroup // Wait group to wait for all async close operations to complete
+	stream          *Stream
+	lazyWrite       bool           // Flag to indicate if lazy write is enabled
+	fileCloseOpt    sync.WaitGroup // Wait group to wait for all async close operations to complete
 }
 
 // Structure defining your config parameters
@@ -101,15 +102,16 @@ type BlockCacheOptions struct {
 }
 
 const (
-	compName               = "block_cache"
-	defaultTimeout         = 120
-	MAX_POOL_USAGE  uint32 = 80
-	MIN_POOL_USAGE  uint32 = 50
-	MIN_PREFETCH           = 5
-	MIN_WRITE_BLOCK        = 3
-	MIN_RANDREAD           = 10
-	MAX_FAIL_CNT           = 3
-	MAX_BLOCKS             = 50000
+	compName                = "block_cache"
+	defaultTimeout          = 120
+	defaultBlockSize        = 16
+	MAX_POOL_USAGE   uint32 = 80
+	MIN_POOL_USAGE   uint32 = 50
+	MIN_PREFETCH            = 5
+	MIN_WRITE_BLOCK         = 3
+	MIN_RANDREAD            = 10
+	MAX_FAIL_CNT            = 3
+	MAX_BLOCKS              = 50000
 )
 
 // Verification to check satisfaction criteria with Component Interface
@@ -170,13 +172,48 @@ func (bc *BlockCache) Stop() error {
 	return nil
 }
 
+// GenConfig : Generate the default config for the component
+func (bc *BlockCache) GenConfig() string {
+	log.Info("BlockCache::Configure : config generation started")
+
+	prefetch := uint32(math.Max((MIN_PREFETCH*2)+1, (float64)(2*runtime.NumCPU())))
+	memSize := uint32(bc.getDefaultMemSize() / _1MB)
+	if (prefetch * defaultBlockSize) > memSize {
+		prefetch = (MIN_PREFETCH * 2) + 1
+	}
+
+	var sb strings.Builder
+	sb.WriteString(fmt.Sprintf("\n%s:", bc.Name()))
+
+	sb.WriteString(fmt.Sprintf("\n  block-size-mb: %v", defaultBlockSize))
+	sb.WriteString(fmt.Sprintf("\n  mem-size-mb: %v", memSize))
+	sb.WriteString(fmt.Sprintf("\n  prefetch: %v", prefetch))
+	sb.WriteString(fmt.Sprintf("\n  parallelism: %v", uint32(3*runtime.NumCPU())))
+
+	var tmpPath string = ""
+	_ = config.UnmarshalKey("tmp-path", &tmpPath)
+	if tmpPath != "" {
+		sb.WriteString(fmt.Sprintf("\n  path: %v", tmpPath))
+		sb.WriteString(fmt.Sprintf("\n  disk-size-mb: %v", bc.getDefaultDiskSize(tmpPath)))
+		sb.WriteString(fmt.Sprintf("\n  disk-timeout-sec: %v", defaultTimeout))
+	}
+
+	return sb.String()
+}
+
 // Configure : Pipeline will call this method after constructor so that you can read config and initialize yourself
 //
 //	Return failure if any config is not valid to exit the process
 func (bc *BlockCache) Configure(_ bool) error {
 	log.Trace("BlockCache::Configure : %s", bc.Name())
+	if common.IsStream {
+		err := bc.stream.Configure(true)
+		if err != nil {
+			log.Err("BlockCache:Stream::Configure : config error [invalid config attributes]")
+			return fmt.Errorf("config error in %s [%s]", bc.Name(), err.Error())
+		}
+	}
 
-	defaultMemSize := false
 	conf := BlockCacheOptions{}
 	err := config.UnmarshalKey(bc.Name(), &conf)
 	if err != nil {
@@ -184,7 +221,7 @@ func (bc *BlockCache) Configure(_ bool) error {
 		return fmt.Errorf("config error in %s [%s]", bc.Name(), err.Error())
 	}
 
-	bc.blockSize = uint64(16) * _1MB
+	bc.blockSize = uint64(defaultBlockSize) * _1MB
 	if config.IsSet(compName + ".block-size-mb") {
 		bc.blockSize = uint64(conf.BlockSize * float64(_1MB))
 	}
@@ -192,15 +229,7 @@ func (bc *BlockCache) Configure(_ bool) error {
 	if config.IsSet(compName + ".mem-size-mb") {
 		bc.memSize = conf.MemSize * _1MB
 	} else {
-		var sysinfo syscall.Sysinfo_t
-		err = syscall.Sysinfo(&sysinfo)
-		if err != nil {
-			log.Err("BlockCache::Configure : config error %s [%s]. Assigning a pre-defined value of 4GB.", bc.Name(), err.Error())
-			bc.memSize = uint64(4192) * _1MB
-		} else {
-			bc.memSize = uint64(0.8 * (float64)(sysinfo.Freeram) * float64(sysinfo.Unit))
-			defaultMemSize = true
-		}
+		bc.memSize = bc.getDefaultMemSize()
 	}
 
 	bc.diskTimeout = defaultTimeout
@@ -212,7 +241,7 @@ func (bc *BlockCache) Configure(_ bool) error {
 	bc.prefetch = uint32(math.Max((MIN_PREFETCH*2)+1, (float64)(2*runtime.NumCPU())))
 	bc.noPrefetch = false
 
-	if defaultMemSize && (uint64(bc.prefetch)*uint64(bc.blockSize)) > bc.memSize {
+	if (!config.IsSet(compName + ".mem-size-mb")) && (uint64(bc.prefetch)*uint64(bc.blockSize)) > bc.memSize {
 		bc.prefetch = (MIN_PREFETCH * 2) + 1
 	}
 
@@ -239,9 +268,20 @@ func (bc *BlockCache) Configure(_ bool) error {
 		bc.workers = conf.Workers
 	}
 
-	bc.tmpPath = ""
-	if conf.TmpPath != "" {
-		bc.tmpPath = common.ExpandPath(conf.TmpPath)
+	bc.tmpPath = common.ExpandPath(conf.TmpPath)
+
+	if bc.tmpPath != "" {
+		//check mnt path is not same as temp path
+		err = config.UnmarshalKey("mount-path", &bc.mntPath)
+		if err != nil {
+			log.Err("BlockCache: config error [unable to obtain Mount Path]")
+			return fmt.Errorf("config error in %s [%s]", bc.Name(), err.Error())
+		}
+
+		if bc.mntPath == bc.tmpPath {
+			log.Err("BlockCache: config error [tmp-path is same as mount path]")
+			return fmt.Errorf("config error in %s error [tmp-path is same as mount path]", bc.Name())
+		}
 
 		// Extract values from 'conf' and store them as you wish here
 		_, err = os.Stat(bc.tmpPath)
@@ -249,31 +289,26 @@ func (bc *BlockCache) Configure(_ bool) error {
 			log.Info("BlockCache: config error [tmp-path does not exist. attempting to create tmp-path.]")
 			err := os.Mkdir(bc.tmpPath, os.FileMode(0755))
 			if err != nil {
-				log.Err("BlockCache: config error creating directory after clean [%s]", err.Error())
+				log.Err("BlockCache: config error creating directory of temp path after clean [%s]", err.Error())
 				return fmt.Errorf("config error in %s [%s]", bc.Name(), err.Error())
 			}
 		}
-		var stat syscall.Statfs_t
-		err = syscall.Statfs(bc.tmpPath, &stat)
-		if err != nil {
-			log.Err("BlockCache::Configure : config error %s [%s]. Assigning a default value of 4GB or if any value is assigned to .disk-size-mb in config.", bc.Name(), err.Error())
-			bc.diskSize = uint64(4192) * _1MB
-		} else {
-			bc.diskSize = uint64(0.8 * float64(stat.Bavail) * float64(stat.Bsize))
-		}
-	}
 
-	if config.IsSet(compName + ".disk-size-mb") {
-		bc.diskSize = conf.DiskSize * _1MB
+		if !common.IsDirectoryEmpty(bc.tmpPath) {
+			log.Err("BlockCache: config error %s directory is not empty", bc.tmpPath)
+			return fmt.Errorf("config error in %s [%s]", bc.Name(), "temp directory not empty")
+		}
+
+		bc.diskSize = bc.getDefaultDiskSize(bc.tmpPath)
+		if config.IsSet(compName + ".disk-size-mb") {
+			bc.diskSize = conf.DiskSize * _1MB
+		}
 	}
 
 	if (uint64(bc.prefetch) * uint64(bc.blockSize)) > bc.memSize {
 		log.Err("BlockCache::Configure : config error [memory limit too low for configured prefetch]")
 		return fmt.Errorf("config error in %s [memory limit too low for configured prefetch]", bc.Name())
 	}
-
-	log.Info("BlockCache::Configure : block size %v, mem size %v, worker %v, prefetch %v, disk path %v, max size %v, disk timeout %v, prefetch-on-open %t, maxDiskUsageHit %v, noPrefetch %v",
-		bc.blockSize, bc.memSize, bc.workers, bc.prefetch, bc.tmpPath, bc.diskSize, bc.diskTimeout, bc.prefetchOnOpen, bc.maxDiskUsageHit, bc.noPrefetch)
 
 	bc.blockPool = NewBlockPool(bc.blockSize, bc.memSize)
 	if bc.blockPool == nil {
@@ -295,7 +330,33 @@ func (bc *BlockCache) Configure(_ bool) error {
 		}
 	}
 
+	log.Crit("BlockCache::Configure : block size %v, mem size %v, worker %v, prefetch %v, disk path %v, max size %v, disk timeout %v, prefetch-on-open %t, maxDiskUsageHit %v, noPrefetch %v",
+		bc.blockSize, bc.memSize, bc.workers, bc.prefetch, bc.tmpPath, bc.diskSize, bc.diskTimeout, bc.prefetchOnOpen, bc.maxDiskUsageHit, bc.noPrefetch)
+
 	return nil
+}
+
+func (bc *BlockCache) getDefaultDiskSize(path string) uint64 {
+	var stat syscall.Statfs_t
+	err := syscall.Statfs(path, &stat)
+	if err != nil {
+		log.Info("BlockCache::getDefaultDiskSize : config error %s [%s]. Assigning a default value of 4GB or if any value is assigned to .disk-size-mb in config.", bc.Name(), err.Error())
+		return uint64(4192) * _1MB
+	}
+
+	return uint64(0.8 * float64(stat.Bavail) * float64(stat.Bsize))
+}
+
+func (bc *BlockCache) getDefaultMemSize() uint64 {
+	var sysinfo syscall.Sysinfo_t
+	err := syscall.Sysinfo(&sysinfo)
+
+	if err != nil {
+		log.Info("BlockCache::getDefaultMemSize : config error %s [%s]. Assigning a pre-defined value of 4GB.", bc.Name(), err.Error())
+		return uint64(4192) * _1MB
+	}
+
+	return uint64(0.8 * (float64)(sysinfo.Freeram) * float64(sysinfo.Unit))
 }
 
 // CreateFile: Create a new file
@@ -333,6 +394,7 @@ func (bc *BlockCache) OpenFile(options internal.OpenFileOptions) (*handlemap.Han
 	handle.Mtime = attr.Mtime
 	handle.Size = attr.Size
 
+	log.Debug("BlockCache::OpenFile : Size of file handle.Size %v", handle.Size)
 	bc.prepareHandleForBlockCache(handle)
 
 	if options.Flags&os.O_TRUNC != 0 || (options.Flags&os.O_WRONLY != 0 && options.Flags&os.O_APPEND == 0) {
@@ -496,6 +558,10 @@ func (bc *BlockCache) closeFileInternal(options internal.CloseFileOptions) error
 	return nil
 }
 
+func (bc *BlockCache) getBlockSize(fileSize uint64, block *Block) uint64 {
+	return min(bc.blockSize, fileSize-block.offset)
+}
+
 // ReadInBuffer: Read the file into a buffer
 func (bc *BlockCache) ReadInBuffer(options internal.ReadInBufferOptions) (int, error) {
 	if options.Offset >= options.Handle.Size {
@@ -522,7 +588,9 @@ func (bc *BlockCache) ReadInBuffer(options internal.ReadInBufferOptions) (int, e
 
 		// Copy data from this block to user buffer
 		readOffset := uint64(options.Offset) - block.offset
-		bytesRead := copy(options.Data[dataRead:], block.data[readOffset:(block.endIndex-block.offset)])
+		blockSize := bc.getBlockSize(uint64(options.Handle.Size), block)
+
+		bytesRead := copy(options.Data[dataRead:], block.data[readOffset:blockSize])
 
 		// Move offset forward in case we need to copy more data
 		options.Offset += int64(bytesRead)
@@ -913,25 +981,31 @@ func (bc *BlockCache) download(item *workItem) {
 				log.Err("BlockCache::download : Failed to open file %s [%s]", fileName, err.Error())
 				_ = os.Remove(localPath)
 			} else {
+				var successfulRead bool = true
 				n, err := f.Read(item.block.data)
 				if err != nil {
 					log.Err("BlockCache::download : Failed to read data from disk cache %s [%s]", fileName, err.Error())
-					f.Close()
+					successfulRead = false
+					_ = os.Remove(localPath)
+				}
+
+				if n != int(bc.blockSize) && item.block.offset+uint64(n) != uint64(item.handle.Size) {
+					log.Err("BlockCache::download : Local data retrieved from disk size mismatch, Expected %v, OnDisk %v, fileSize %v", bc.getBlockSize(uint64(item.handle.Size), item.block), n, item.handle.Size)
+					successfulRead = false
 					_ = os.Remove(localPath)
 				}
 
 				f.Close()
 				// We have read the data from disk so there is no need to go over network
 				// Just mark the block that download is complete
-
-				item.block.endIndex = item.block.offset + uint64(n)
-				item.block.Ready(BlockStatusDownloaded)
-				return
+				if successfulRead {
+					item.block.Ready(BlockStatusDownloaded)
+					return
+				}
 			}
 		}
 	}
 
-	item.block.endIndex = item.block.offset
 	// If file does not exists then download the block from the container
 	n, err := bc.NextComponent().ReadInBuffer(internal.ReadInBufferOptions{
 		Handle: item.handle,
@@ -960,8 +1034,6 @@ func (bc *BlockCache) download(item *workItem) {
 		bc.threadPool.Schedule(false, item)
 		return
 	}
-
-	item.block.endIndex = item.block.offset + uint64(n)
 
 	if bc.tmpPath != "" {
 		err := os.MkdirAll(filepath.Dir(localPath), 0777)
@@ -1020,10 +1092,6 @@ func (bc *BlockCache) WriteFile(options internal.WriteFileOptions) (int, error) 
 		// Move offset forward in case we need to copy more data
 		options.Offset += int64(bytesWritten)
 		dataWritten += bytesWritten
-
-		if block.endIndex < uint64(options.Offset) {
-			block.endIndex = uint64(options.Offset)
-		}
 
 		if options.Handle.Size < options.Offset {
 			options.Handle.Size = options.Offset
@@ -1247,28 +1315,15 @@ func shouldCommitAndDownload(blockID int64, handle *handlemap.Handle) (bool, boo
 
 // lineupUpload : Create a work item and schedule the upload
 func (bc *BlockCache) lineupUpload(handle *handlemap.Handle, block *Block, listMap map[int64]*blockInfo) {
-	// if a block has data less than block size and is not the last block,
-	// add null at the end and upload the full block
-	// bc.printCooking(handle)
-	if block.endIndex < uint64(handle.Size) {
-		log.Debug("BlockCache::lineupUpload : Appending null for block %v, size %v for %v=>%s", block.id, (block.endIndex - block.offset), handle.ID, handle.Path)
-		block.endIndex = block.offset + bc.blockSize
-	} else if block.endIndex == uint64(handle.Size) {
-		// TODO: random write scenario where this block is not the last block
-		log.Debug("BlockCache::lineupUpload : Last block %v, size %v for %v=>%s", block.id, (block.endIndex - block.offset), handle.ID, handle.Path)
-	}
 
-	// id := listMap[block.id]
-	// if id == "" {
 	id := base64.StdEncoding.EncodeToString(common.NewUUIDWithLength(16))
 	listMap[block.id] = &blockInfo{
 		id:        id,
 		committed: false,
-		size:      block.endIndex - block.offset,
+		size:      bc.getBlockSize(uint64(handle.Size), block),
 	}
-	//}
 
-	log.Debug("BlockCache::lineupUpload : block %v, size %v for %v=>%s, blockId %v", block.id, (block.endIndex - block.offset), handle.ID, handle.Path, id)
+	log.Debug("BlockCache::lineupUpload : block %v, size %v for %v=>%s, blockId %v", block.id, bc.getBlockSize(uint64(handle.Size), block), handle.ID, handle.Path, id)
 	item := &workItem{
 		handle:   handle,
 		block:    block,
@@ -1277,8 +1332,6 @@ func (bc *BlockCache) lineupUpload(handle *handlemap.Handle, block *Block, listM
 		upload:   true,
 		blockId:  id,
 	}
-
-	// log.Debug("BlockCache::lineupUpload : Upload block %v=>%s (index %v, offset %v, data %v)", handle.ID, handle.Path, block.id, block.offset, (block.endIndex - block.offset))
 
 	block.Uploading()
 	block.flags.Clear(BlockFlagFailed)
@@ -1346,12 +1399,11 @@ func (bc *BlockCache) upload(item *workItem) {
 	flock := bc.fileLocks.Get(fileName)
 	flock.Lock()
 	defer flock.Unlock()
-	// log.Debug("BlockCache::Upload : block %v, size %v for %v=>%s, blockId %v", item.block.id, (item.block.endIndex - item.block.offset), item.handle.ID, item.handle.Path, item.blockId)
-
+	blockSize := bc.getBlockSize(uint64(item.handle.Size), item.block)
 	// This block is updated so we need to stage it now
 	err := bc.NextComponent().StageData(internal.StageDataOptions{
 		Name:   item.handle.Path,
-		Data:   item.block.data[0 : item.block.endIndex-item.block.offset],
+		Data:   item.block.data[0:blockSize],
 		Offset: uint64(item.block.offset),
 		Id:     item.blockId})
 	if err != nil {
@@ -1383,7 +1435,7 @@ func (bc *BlockCache) upload(item *workItem) {
 		// Dump this block to local disk cache
 		f, err := os.Create(localPath)
 		if err == nil {
-			_, err := f.Write(item.block.data[0 : item.block.endIndex-item.block.offset])
+			_, err := f.Write(item.block.data[0:blockSize])
 			if err != nil {
 				log.Err("BlockCache::upload : Failed to write %s to disk [%v]", localPath, err.Error())
 				_ = os.Remove(localPath)